+++ src/common/components/Tip.vue
@@ -6,7 +6,7 @@ defineProps<{ tip: string }>()
     <span
         v-tippy="tip"
         :class="[
-            'border rounded-full w-4 h-4 text-center ml-1 cursor-pointer inline-flex items-center',
+            'border rounded-full size-4 text-center ml-1 cursor-pointer inline-flex items-center',
             'justify-center border-gray-500 text-xs font-heading',
         ]"
         >?</span