+++ src/popup/components/Navbar/Sidebar/SidebarLink.vue
@@ -15,7 +15,7 @@ defineProps<Props>()
         target="_blank"
         class="flex items-center gap-3 bg-page hover:bg-page-hover transition-colors px-4 py-2"
     >
-        <component :is="icon" class="w-5 h-5" />
+        <component :is="icon" class="size-5" />
         <span class="text-sm">{{ label }}</span>
     </a>
 </template>