+++ src/popup/components/Views/MainView/NewGroup/Buttons.vue
@@ -24,7 +24,7 @@ async function askForGroupName(isPrivate: boolean) {
             @click="askForGroupName(false)"
             class="w-full bg-primary hover:bg-primary-hover"
         >
-            <PlusCircleIcon class="w-6 h-6" />
+            <PlusCircleIcon class="size-6" />
             <span>{{ trans('new_group') }}</span>
         </NewGroupButton>
 
@@ -33,7 +33,7 @@ async function askForGroupName(isPrivate: boolean) {
             @click="askForGroupName(true)"
             class="w-24 bg-private hover:bg-private-hover"
         >
-            <ShieldCheckIcon class="w-8 h-8" />
+            <ShieldCheckIcon class="size-8" />
         </NewGroupButton>
     </div>
 </template>