+++ src/popup/components/Views/SelectTabsView/TabItem.vue
@@ -16,13 +16,13 @@ defineProps<{
     <LinkElement @click="emit('toggle')" :isSelected :link>
         <label
             :class="[
-                'w-4 h-4 shrink-0 border border-border flex items-center justify-center',
+                'size-4 shrink-0 border border-border flex items-center justify-center',
                 'cursor-pointer rounded-full',
             ]"
         >
             <input @change="emit('toggle')" type="checkbox" class="hidden" />
 
-            <div v-if="isSelected" class="w-3 h-3 bg-private rounded-full"></div>
+            <div v-if="isSelected" class="size-3 bg-private rounded-full"></div>
         </label>
     </LinkElement>
 </template>